Split config parsing and env parsing
